Create shared module
