+++ app/build.gradle
@@ -55,6 +55,7 @@ dependencies {
     implementation project(":users")
     implementation project(":registration")
     implementation project(":tech_groups")
+    implementation project(":shared")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
