+++ calendar/build.gradle
@@ -4,11 +4,10 @@ dependencies {
 
     implementation project(":ui")
     implementation project(":repository")
+    implementation project(":shared")
 
     def nav_version = "2.3.4"
 
-    implementation project(":ui")
-    implementation project(":repository")
 
     // gson
     implementation 'com.google.code.gson:gson:2.8.5'
