+++ calendar/src/main/java/com/intive/calendar/components/MonthViewComponents.kt
@@ -23,6 +23,8 @@ import androidx.navigation.NavController
 import com.intive.calendar.R
 import com.intive.calendar.screens.CalendarHeader
 import com.intive.calendar.utils.*
+import com.intive.shared.getDateString
+import com.intive.shared.getFullDateString
 import java.util.*
 
 
@@ -99,12 +101,7 @@ fun CalendarGrid(
 
                             val eventBundle = EventBundle(
                                 id = monthEvents[index].events!![0].id,
-                                date = "${weekDaysCalendarClass[(items[it] as Calendar)[Calendar.DAY_OF_WEEK]]}, ${
-                                    getDateString(
-                                        (items[it] as Calendar),
-                                        "."
-                                    )
-                                }",
+                                date = getFullDateString(items[it] as Calendar),
                                 time = "${monthEvents[index].events!![0].timeStart} - ${monthEvents[index].events!![0].timeEnd}",
                                 name = monthEvents[index].events!![0].name,
                                 inviteResponse = monthEvents[index].events!![0].inviteResponse,
@@ -131,12 +128,7 @@ fun CalendarGrid(
                             val isDayActive = !((items[it] as Calendar).before(Calendar.getInstance()))
 
                             val dayBundle = DayBundle(
-                                date = "${weekDaysCalendarClass[(items[it] as Calendar)[Calendar.DAY_OF_WEEK]]}, ${
-                                    getDateString(
-                                        (items[it] as Calendar),
-                                        "."
-                                    )
-                                }",
+                                date = getFullDateString(items[it] as Calendar),
                                 events = monthEvents[index].events!!,
                                 active = isDayActive
                             )
