+++ calendar/src/main/java/com/intive/calendar/screens/AddEventScreen.kt
@@ -19,10 +19,10 @@ import com.intive.calendar.R
 import com.intive.calendar.viewmodels.AddEventViewModel
 import java.util.*
 import com.intive.calendar.components.*
-import com.intive.calendar.utils.getDateString
 import com.intive.ui.components.TitleText
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.ui.platform.LocalFocusManager
+import com.intive.shared.getDateString
 import com.intive.ui.components.CheckBoxesList
 import com.intive.ui.components.PrimaryButton
