+++ calendar/src/main/java/com/intive/calendar/utils/Constants.kt
@@ -2,7 +2,6 @@ package com.intive.calendar.utils
 
 val calendarHeader = listOf("Pn", "Wt", "Śr", "Cz", "Pt", "Sb", "Nd")
 val weekDays = listOf("Poniedziałek", "Wtorek", "Środa", "Czwartek", "Piątek", "Sobota", "Niedziela")
-val weekDaysCalendarClass = listOf("", "Niedziela", "Poniedziałek", "Wtorek", "Środa", "Czwartek", "Piątek", "Sobota")
 const val userId: Long = 1
 const val eventBundleKey = "event"
 const val dayBundleKey = "day"
