+++ settings.gradle
@@ -8,3 +8,4 @@ include ':calendar'
 include ':audit'
 include ':repository'
 include ':tech_groups'
+include ':shared'
