+++ shared/build.gradle
@@ -1,7 +1,11 @@
 apply from: "$rootProject.projectDir/shared.gradle"
 
 dependencies {
+
     def nav_version = "2.3.4"
+
+    // gson
+    implementation 'com.google.code.gson:gson:2.8.5'
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
@@ -57,7 +61,9 @@ dependencies {
     // Koin for Jetpack Compose (unstable version)
     implementation "io.insert-koin:koin-androidx-compose:$koin_version"
 
-    implementation project(":ui")
-    implementation project(":repository")
-    implementation project(":shared")
+    //Paging
+    implementation "androidx.paging:paging-runtime-ktx:3.0.0-beta03"
+    implementation "androidx.paging:paging-compose:1.0.0-alpha08"
 }
+
+
