+++ shared/src/main/java/com/intive/shared/Constants.kt
@@ -1,0 +1,3 @@
+package com.intive.shared
+
+val weekDaysCalendarClass = listOf("", "Niedziela", "Poniedziałek", "Wtorek", "Środa", "Czwartek", "Piątek", "Sobota")
